# public/swagger.yaml
openapi: 3.0.3
info:
  title: Dentista Now API
  description: |
    API completa para gerenciamento de barbearia incluindo agendamentos, barbeiros, clientes, produtos, planos e webhooks.

    ## Autenticação
    A maioria dos endpoints requer autenticação via Bearer Token no header:
    ```
    Authorization: Bearer SEU_TOKEN_AQUI
    ```

    ## Formato de Datas
    - Datas são no formato brasileiro: DD/MM/YYYY ou YYYY-MM-DD
    - Horários são no formato: HH:MM (24h)

    ## Webhooks
    O sistema suporta dois tipos de webhooks:
    1. **Webhook Global**: Configurado na tabela `configuracoes`, dispara para todos os eventos
    2. **Webhooks de Barbeiros**: Webhooks personalizados por profissional

  version: 3.0.0
  contact:
    name: Dentista Now
    url: https://dentista-now.vercel.app
servers:
  - url: https://dentista-now.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticação da API
...
paths:
...
  /admin/estatisticas:
    get:
      tags:
        - Admin
      summary: Estatísticas gerais do sistema
      description: |
        Retorna estatísticas gerais da barbearia para o período especificado.

        **Períodos disponíveis:**
        - hoje
        - semana
        - mes (padrão)
        - ano
      security:
        - BearerAuth: []
      parameters:
        - name: periodo
          in: query
          schema:
            type: string
            default: mes
            enum: [hoje, semana, mes, ano]
          description: Período das estatísticas
        - name: data_inicio
          in: query
          schema:
            type: string
          description: Data de início customizada (DD-MM-YYYY)
        - name: data_fim
          in: query
          schema:
            type: string
          description: Data fim customizada (DD-MM-YYYY)
      responses:
        '200':
          description: Estatísticas retornadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_agendamentos:
                    type: integer
                  total_clientes:
                    type: integer
                  faturamento_mensal:
                    type: number
                  agendamentos_hoje:
                    type: integer

  # =====================
  # SYNC E DEBUG
  # =====================
...